app2.0/backend/requirements.txt
absl-py==0.11.0
astroid==2.4.2
astunparse==1.6.3
attrs==20.3.0
bcrypt==3.2.0
blis==0.7.4
bson==0.5.10
cachetools==4.2.1
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.5
cymem==2.0.5
dnspython==1.16.0
docx2txt==0.8
en-core-web-sm==2.3.1
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==3.25.0
Flask-PyMongo==2.3.0
flatbuffers==1.12
gast==0.3.3
gensim==3.8.3
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
motor==2.3.1
murmurhash==1.0.5
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
openpyxl==3.0.6
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.2
pathy==0.4.0
pdfminer==20191125
pdfminer.six==20201018
pdfminer3==2018.12.3.0
Pillow==8.1.0
plac==1.1.3
preshed==3.0.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.3
PyJWT==1.7.1
pylint==2.6.2
pymongo==3.11.3
pyparsing==2.4.7
pyresparser==1.0.6
pyrsistent==0.17.3
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.0
sentence-transformers==0.4.0
sentencepiece==0.1.95
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.3.0
spacy==2.3.5
spacy-legacy==3.0.1
srsly==1.0.5
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
thinc==7.4.5
threadpoolctl==2.1.0
tokenizers==0.10.1
toml==0.10.2
torch==1.7.1
tqdm==4.56.2
transformers==4.3.3
typed-ast==1.4.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
